+++ src/table_one.jl
@@ -88,7 +88,7 @@ function Analysis(s::Symbol, df::DataFrames.DataFrame)
     Analysis(s, default_analysis(df[!, s]), string(s))
 end
 
-function Analysis(p::Pair{Symbol, <:Any}, df::DataFrames.DataFrame)
+function Analysis(p::Pair{<:Union{Symbol,String}, <:Any}, df::DataFrames.DataFrame)
     sym, rest = p
     Analysis(sym, rest, df)
 end
@@ -112,6 +112,10 @@ end
 function Analysis(sym::Symbol, p::Pair, df::DataFrames.DataFrame)
     funcs, name = p
     Analysis(sym, funcs, name, df)
+end
+
+function Analysis(sym::String, args...)
+    Analysis(Symbol(sym), args...)
 end
 
 make_analyses(v::AbstractVector, df::DataFrame) = map(x -> Analysis(x, df), v)
@@ -226,7 +230,7 @@ can be stratified by one, or more, variables using the `groupby` keyword.
   - `tests`: A `NamedTuple` of hypothesis test types to use for `categorical`, `nonnormal`, `minmax`, and `normal` variables.
   - `combine`: An object from `MultipleTesting` to use when combining p-values.
   - `show_total`: Display the total column summary. Default is `true`.
-  - `group_totals`: A group `Symbol` or vector of symbols specifying for which group levels totals should be added. Any group levels but the topmost can be chosen (the topmost being already handled by the `show_total` option). Default is `Symbol[]`.
+  - `group_totals`: A group `Symbol` or `String` or vector of symbols/strings specifying for which group levels totals should be added. Any group levels but the topmost can be chosen (the topmost being already handled by the `show_total` option). Default is `Symbol[]`.
   - `total_name`: The name for all total columns. Default is `"Total"`.
   - `show_n`: Display the number of rows for each group key next to its label.
   - `show_pvalues`: Display the `P-Value` column. Default is `false`.
@@ -292,8 +296,9 @@ function table_one(
 
     groupsymbols = [g.symbol for g in groups]
 
-    _group_totals(a::AbstractVector{Symbol}) = collect(a)
+    _group_totals(a::AbstractVector{<:Union{String,Symbol}}) = Symbol.(a)
     _group_totals(s::Symbol) = [s]
+    _group_totals(s::String) = [Symbol(s)]
     group_totals = _group_totals(group_totals) 
     if !isempty(groupsymbols) && first(groupsymbols) in group_totals
         throw(ArgumentError("Cannot show totals for topmost group $(repr(first(groupsymbols))) as it would be equivalent to the `show_total` option. Grouping is $groupsymbols"))